Move test_figshare_groups() into test_commons
